requirements.txt
PyQt5
pycrypto
requests